Changed NodeId from u64 to u32

+++ common/cosmwasm-smart-contracts/mixnet-contract/src/types.rs
@@ -19,7 +19,7 @@ pub type SphinxKey = String;
 pub type SphinxKeyRef<'a> = &'a str;
 pub type EpochId = u32;
 pub type IntervalId = u32;
-pub type NodeId = u64;
+pub type NodeId = u32;
 pub type EpochEventId = u32;
 pub type IntervalEventId = u32;
